meme/requirements.txt
solana>=0.32.0
solders>=0.21.0
httpx>=0.27.0
pyyaml>=6.0
pydantic>=2.5.0
rich>=13.7.0
tabulate>=0.9.0
aiohttp>=3.9.0
base58>=2.1.0
